requirements.txt
PyYAML~=6.0.2
Wand~=0.6.13
requests~=2.32.3